removing review.py
